PM / devfreq: Make the monitor workqueue high priority
Running the devfreq monitor in a normal-priority workqueue can result
in the devfreq mutex lock being held for a while when the monitor
worker sleeps inside update_devfreq(). Freeing up the devfreq mutex
lock is important to ensure timely devfreq boosts, so run the monitor
workqueue in a high-priority workqueue to reduce the lock duration.

Change-Id: Id7ccdbb4d2b566266faf0d83a32e20643dbc11fc
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>